Update hello.sh
start opensuse support

+++ hello.sh
@@ -5,10 +5,11 @@ if [ -e /etc/debian_version ]; then
     sleep 5;
   done
   apt-get update && apt-get -y upgrade && apt-get -y install default-jre git
-else
+elif [ -e /etc/centos_version ]; then
   yum -y install git java-1.8.0-openjdk
   sed -i -e '/secure_path/ s[=.*[&:/usr/local/share/fusionforge/bin:/usr/local/bin[' /etc/sudoers
-  sed -i -e 's/^.*env_keep.*$//' /etc/sudoers
+else
+  zypper --non-interactive install git java-10-openjdk
 fi
 #force sudoers to handle waagent ...
 echo 'slave ALL=(ALL) NOPASSWD: ALL' > /etc/sudoers.d/zzz-slave
